[fix] [broker] delete topic failed if disabled system topic (#19735)
Motivation: After PR #18823, The cmd delete topic will fail if disabled the feature system topic.
Modifications: do not delete the system policy if disabled the feature system topic

(cherry picked from commit 401fb055ed428b7f3e33a49cea78b25b04f7448e)

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/service/persistent/PersistentTopic.java
@@ -1183,7 +1183,8 @@ public class PersistentTopic extends AbstractTopic
                         deleteTopicAuthenticationFuture.thenCompose(ignore -> deleteSchema ? deleteSchema() :
                                         CompletableFuture.completedFuture(null))
                                 .thenCompose(ignore -> {
-                                    if (!isTopicPoliciesSystemTopic(topic)) {
+                                    if (!isTopicPoliciesSystemTopic(topic)
+                                            && brokerService.getPulsar().getConfiguration().isSystemTopicEnabled()) {
                                         return deleteTopicPolicies();
                                     } else {
                                         return CompletableFuture.completedFuture(null);
